modulate/requirements.txt
aiohttp>=3.9.0
sounddevice>=0.4.6
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
